ipiranga-python-integration-factory/ipiranga-python-integration-factory/templates/rest/app/domain/backend_payload.py
from __future__ import annotations

from typing import Any

from pydantic import BaseModel


class BackendPayload(BaseModel):
    """Outbound payload model for {{ service_name }}."""

    data: dict[str, Any]
