ipiranga-python-integration-factory/ipiranga-python-integration-factory/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ---------------------------------------------------------------------------
# Project metadata
# ---------------------------------------------------------------------------
[project]
name = "ipiranga-python-integration-factory"
version = "0.1.0"
description = "CLI generator for Ipiranga integration microservices"
readme = "README.md"
license = { text = "Proprietary" }
requires-python = ">=3.12"

authors = [
    { name = "Ipiranga Platform Engineering" },
]

keywords = ["generator", "microservices", "fastapi", "integration", "ipiranga"]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

# ---------------------------------------------------------------------------
# Runtime dependencies
# ---------------------------------------------------------------------------
dependencies = [
    "PyYAML>=6.0",
]

# ---------------------------------------------------------------------------
# Optional / development dependencies
# ---------------------------------------------------------------------------
[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

# ---------------------------------------------------------------------------
# CLI entry point
# ---------------------------------------------------------------------------
[project.scripts]
integration-factory = "generator.cli:run"

# ---------------------------------------------------------------------------
# Tool configuration
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["generator"]
